Update the permissions to be compatible with changes from Transport Layer migration
