+++ plugins/content-manager/src/main/java/com/wazuh/contentmanager/rest/service/RestPutPolicyAction.java
@@ -17,24 +17,29 @@
 package com.wazuh.contentmanager.rest.service;
 
 import com.fasterxml.jackson.databind.JsonNode;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.fasterxml.jackson.databind.node.ObjectNode;
 
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
+import org.opensearch.action.index.IndexResponse;
 import org.opensearch.core.rest.RestStatus;
 import org.opensearch.rest.BaseRestHandler;
 import org.opensearch.rest.RestRequest;
-import org.opensearch.rest.action.RestResponseListener;
 import org.opensearch.transport.client.node.NodeClient;
 
-import java.util.List;
+import java.io.IOException;
+import java.time.Instant;
+import java.util.*;
 
-import com.wazuh.contentmanager.action.PutPolicyAction;
-import com.wazuh.contentmanager.action.PutPolicyRequest;
-import com.wazuh.contentmanager.action.PutPolicyResponse;
+import com.wazuh.contentmanager.cti.catalog.index.ContentIndex;
+import com.wazuh.contentmanager.cti.catalog.model.Policy;
+import com.wazuh.contentmanager.cti.catalog.model.Resource;
 import com.wazuh.contentmanager.cti.catalog.model.Space;
 import com.wazuh.contentmanager.cti.catalog.service.SpaceService;
 import com.wazuh.contentmanager.engine.service.EngineService;
 import com.wazuh.contentmanager.rest.model.RestResponse;
+import com.wazuh.contentmanager.rest.utils.PayloadValidations;
 import com.wazuh.contentmanager.settings.PluginSettings;
 import com.wazuh.contentmanager.utils.Constants;
 
@@ -44,10 +49,7 @@ import static org.opensearch.rest.RestRequest.Method.PUT;
  * REST handler for updating policy resources on the Wazuh Engine.
  *
  * <p>This endpoint handles PUT requests to update policy configurations in the draft or standard
- * space. The actual write is delegated to {@link PutPolicyAction} so the security plugin enforces
- * authorization, and the {@code plugins.content_manager.sensitive_config.locked} setting can block
- * the operation. When the transport action reports a standard-space hash change, this handler
- * reloads the standard space into the Engine (best effort).
+ * space. The policy defines the root decoder and integrations list for content processing.
  */
 public class RestPutPolicyAction extends BaseRestHandler {
     private static final Logger log = LogManager.getLogger(RestPutPolicyAction.class);
@@ -55,24 +57,45 @@ public class RestPutPolicyAction extends BaseRestHandler {
 
     private SpaceService spaceService;
     private EngineService engineService;
+    private NodeClient client;
+    private PayloadValidations payloadValidations;
+
+    private static final ObjectMapper mapper = new ObjectMapper();
 
     /**
      * Constructs a new RestPutPolicyAction handler.
      *
-     * @param spaceService The space service instance used to build the Engine payload.
+     * @param spaceService The space service instance to fetch policies.
      * @param engineService The engine service instance for loading policies into the Engine.
      */
     public RestPutPolicyAction(SpaceService spaceService, EngineService engineService) {
         this.spaceService = spaceService;
         this.engineService = engineService;
+        this.payloadValidations = new PayloadValidations();
     }
 
     /**
-     * Setter for the space service, used in tests.
+     * Constructs a new RestPutPolicyAction handler with explicit NodeClient (for testing or DI).
      *
-     * @param spaceService the space service instance to set
+     * @param spaceService The space service instance to fetch policies.
+     * @param engineService The engine service instance for loading policies into the Engine.
+     * @param client The NodeClient to use for index operations. TODO should not be required to pass
+     *     the client
      */
-    public void setSpaceService(SpaceService spaceService) {
+    public RestPutPolicyAction(
+            SpaceService spaceService, EngineService engineService, NodeClient client) {
+        this.spaceService = spaceService;
+        this.engineService = engineService;
+        this.client = client;
+        this.payloadValidations = new PayloadValidations();
+    }
+
+    /**
+     * Setter for the policy hash service, used in tests.
+     *
+     * @param spaceService the policy hash service to set
+     */
+    public void setPolicyHashService(SpaceService spaceService) {
         this.spaceService = spaceService;
     }
 
@@ -83,6 +106,15 @@ public class RestPutPolicyAction extends BaseRestHandler {
      */
     public void setEngineService(EngineService engineService) {
         this.engineService = engineService;
+    }
+
+    /**
+     * Setter for the payload validations, used in tests.
+     *
+     * @param payloadValidations the payload validations instance to set
+     */
+    public void setPayloadValidations(PayloadValidations payloadValidations) {
+        this.payloadValidations = payloadValidations;
     }
 
     /** Return a short identifier for this handler. */
@@ -102,37 +134,391 @@ public class RestPutPolicyAction extends BaseRestHandler {
     }
 
     /**
-     * Delegates the policy update to {@link PutPolicyAction}, then reloads the Engine when required.
+     * Prepares the request by returning a consumer that executes the policy update operation.
      *
      * @param request the incoming REST request containing the policy payload
-     * @param client the node client used to execute the transport action
-     * @return a consumer that sends the policy update response
+     * @param client the node client for index operations
+     * @return a consumer that executes the policy update operation
      */
     @Override
     protected RestChannelConsumer prepareRequest(RestRequest request, NodeClient client) {
-        String space = request.param(Constants.KEY_SPACE);
-        String payload = request.hasContent() ? request.content().utf8ToString() : null;
-        PutPolicyRequest putPolicyRequest = new PutPolicyRequest(space, payload);
-        return channel ->
-                client.execute(
-                        PutPolicyAction.INSTANCE,
-                        putPolicyRequest,
-                        new RestResponseListener<PutPolicyResponse>(channel) {
-                            @Override
-                            public org.opensearch.rest.RestResponse buildResponse(PutPolicyResponse response) {
-                                if (response.shouldReloadEngine()) {
-                                    loadStandardSpaceIntoEngine();
-                                }
-                                return new RestResponse(response.getMessage(), response.getStatus().getStatus())
-                                        .toBytesRestResponse();
-                            }
-                        });
+        this.client = client;
+        this.spaceService = new SpaceService(client);
+        RestResponse response = this.handleRequest(request);
+        return channel -> channel.sendResponse(response.toBytesRestResponse());
     }
 
     /**
-     * Builds the engine payload for the standard space and loads it into the Engine. Best effort:
-     * failures are logged but do not affect the policy update response.
+     * Handles the policy update request by validating the payload and storing the policy.
+     *
+     * <p>This method performs the following validations:
+     *
+     * <ol>
+     *   <li>Checks that the engine service is available
+     *   <li>Verifies that the request contains a JSON payload
+     *   <li>Parses and validates the Policy JSON structure
+     *   <li>Validates the space path parameter
+     * </ol>
+     *
+     * @param request incoming REST request containing the policy data
+     * @return a RestResponse describing the outcome of the operation
      */
+    public RestResponse handleRequest(RestRequest request) {
+        // 1. Check request's payload exists
+        if (request == null || !request.hasContent()) {
+            return new RestResponse(
+                    Constants.E_400_INVALID_REQUEST_BODY, RestStatus.BAD_REQUEST.getStatus());
+        }
+        try {
+            // Extract and validate space parameter
+            String spaceName = request.param(Constants.KEY_SPACE);
+            if (!Space.DRAFT.equals(spaceName) && !Space.STANDARD.equals(spaceName)) {
+                return new RestResponse(
+                        String.format(
+                                Locale.ROOT,
+                                Constants.E_400_RESOURCE_SPACE_MISMATCH,
+                                Space.DRAFT + ", " + Space.STANDARD),
+                        RestStatus.BAD_REQUEST.getStatus());
+            }
+
+            // 2. Validate request content
+            JsonNode jsonContent;
+            try {
+                jsonContent = mapper.readTree(request.content().utf8ToString());
+            } catch (IOException e) {
+                return new RestResponse(
+                        Constants.E_400_INVALID_REQUEST_BODY, RestStatus.BAD_REQUEST.getStatus());
+            }
+
+            // Validate "resource"
+            if (!jsonContent.has(Constants.KEY_RESOURCE)) {
+                return new RestResponse(
+                        String.format(Locale.ROOT, Constants.E_400_MISSING_FIELD, Constants.KEY_RESOURCE),
+                        RestStatus.BAD_REQUEST.getStatus());
+            }
+            JsonNode resource = jsonContent.get(Constants.KEY_RESOURCE);
+            log.debug(Constants.D_LOG_OPERATION, "Updating", Constants.KEY_POLICY, resource);
+            Policy policy;
+            try {
+                policy = mapper.readValue(resource.toString(), Policy.class);
+            } catch (IOException e) {
+                return new RestResponse(
+                        Constants.E_400_INVALID_REQUEST_BODY, RestStatus.BAD_REQUEST.getStatus());
+            }
+
+            // Normalize empty root_decoder to null so it is omitted from serialization
+            if (policy.getRootDecoder() != null && policy.getRootDecoder().isEmpty()) {
+                policy.setRootDecoder(null);
+            }
+
+            // Validate required Policy fields
+            List<String> missingFields = new ArrayList<>();
+            if (policy.getEnabled() == null) {
+                missingFields.add(Constants.KEY_ENABLED);
+            }
+            if (policy.getIndexUnclassifiedEvents() == null) {
+                missingFields.add("index_unclassified_events");
+            }
+            if (policy.getIndexDiscardedEvents() == null) {
+                missingFields.add("index_discarded_events");
+            }
+
+            // Draft space requires additional fields
+            if (Space.DRAFT.equals(spaceName)) {
+                if (policy.getAuthor() == null || policy.getAuthor().isEmpty()) {
+                    missingFields.add(Constants.KEY_AUTHOR);
+                }
+                if (policy.getDescription() == null || policy.getDescription().isEmpty()) {
+                    missingFields.add(Constants.KEY_DESCRIPTION);
+                }
+                if (policy.getDocumentation() == null) {
+                    missingFields.add("documentation");
+                }
+                if (policy.getReferences() == null) {
+                    missingFields.add("references");
+                }
+            }
+
+            if (!missingFields.isEmpty()) {
+                return new RestResponse(
+                        String.format(
+                                Locale.ROOT, Constants.E_400_MISSING_FIELD, String.join(", ", missingFields)),
+                        RestStatus.BAD_REQUEST.getStatus());
+            }
+
+            Set<String> knownEnrichmentTypes = this.spaceService.getKnownEnrichmentTypes();
+
+            // Validate enrichments: only allowed values, no duplicates
+            RestResponse enrichmentsValidationError =
+                    this.payloadValidations.validateEnrichments(
+                            policy.getEnrichments(), knownEnrichmentTypes);
+            if (enrichmentsValidationError != null) {
+                return enrichmentsValidationError;
+            }
+
+            // 3. Update policy based on target space
+            String policyId;
+            if (Space.STANDARD.equals(spaceName)) {
+                policyId = this.updateStandardPolicy(policy);
+            } else {
+                policyId = this.updatePolicy(policy);
+            }
+
+            // Regenerate space hash because space composition changed
+            Set<String> changedSpaces = this.spaceService.calculateAndUpdate(List.of(spaceName));
+
+            // Load the standard space into the Engine only if its hash changed
+            if (changedSpaces.contains(Space.STANDARD.toString())) {
+                this.loadStandardSpaceIntoEngine();
+            }
+
+            return new RestResponse(policyId, RestStatus.OK.getStatus());
+        } catch (IllegalArgumentException e) {
+            log.warn(Constants.W_LOG_VALIDATION_FAILED, e.getMessage());
+            return new RestResponse(
+                    Constants.E_400_INVALID_REQUEST_BODY + " " + e.getMessage(),
+                    RestStatus.BAD_REQUEST.getStatus());
+        } catch (Exception e) {
+            log.error(
+                    Constants.E_LOG_OPERATION_FAILED, "updating", Constants.KEY_POLICY, e.getMessage(), e);
+            return new RestResponse(
+                    Constants.E_500_INTERNAL_SERVER_ERROR + " " + e.getMessage(),
+                    RestStatus.INTERNAL_SERVER_ERROR.getStatus());
+        }
+    }
+
+    /**
+     * Stores or updates the policy in the standard space.
+     *
+     * <p>Only the following fields from the incoming policy are applied: enrichments, filters,
+     * enabled, index_unclassified_events, and index_discarded_events. All other fields are preserved
+     * from the existing standard policy document.
+     *
+     * @param incomingPolicy the incoming policy containing the fields to update
+     * @return the document ID of the persisted policy
+     * @throws IOException if storage fails
+     * @throws IllegalStateException if the standard policy document is not found
+     */
+    @SuppressWarnings("unchecked")
+    private String updateStandardPolicy(Policy incomingPolicy)
+            throws IOException, IllegalStateException {
+        // Get current standard policy
+        Map<String, Object> currentPolicy = this.spaceService.getPolicy(Space.STANDARD.toString());
+        Map<String, Object> currentPolicyDoc =
+                (Map<String, Object>) currentPolicy.get(Constants.KEY_DOCUMENT);
+        if (currentPolicyDoc == null) {
+            throw new IllegalStateException(
+                    Constants.E_500_INTERNAL_SERVER_ERROR + " Policy document not found in standard space.");
+        }
+
+        // Validate filters: allow reordering but prevent addition/removal
+        List<String> currentFilters =
+                (List<String>)
+                        currentPolicyDoc.getOrDefault(Constants.KEY_FILTERS, Collections.emptyList());
+        List<String> newFilters = incomingPolicy.getFilters();
+
+        RestResponse filtersValidationError =
+                this.payloadValidations.validateListEquality(
+                        currentFilters, newFilters, Constants.KEY_FILTERS);
+        if (filtersValidationError != null) {
+            throw new IllegalArgumentException(filtersValidationError.getMessage());
+        }
+
+        // Safekeep all unmodifiable values from the existing standard policy
+        String docId = currentPolicyDoc.getOrDefault(Constants.KEY_ID, "").toString();
+
+        // Extract metadata from existing policy (now nested under "metadata", fallback to root)
+        @SuppressWarnings("unchecked")
+        Map<String, Object> existingMetadata =
+                (Map<String, Object>)
+                        currentPolicyDoc.getOrDefault(Constants.KEY_METADATA, Collections.emptyMap());
+
+        Object dateObj = existingMetadata.get(Constants.KEY_DATE);
+        if (dateObj == null) dateObj = currentPolicyDoc.get(Constants.KEY_DATE);
+        String docCreationDate = dateObj != null ? dateObj.toString() : "";
+
+        // Build a policy preserving existing fields, only overriding the 5 allowed fields
+        Policy mergedPolicy = new Policy();
+        mergedPolicy.setId(docId);
+        mergedPolicy.setDate(docCreationDate);
+        mergedPolicy.setModified(Instant.now().toString());
+
+        Object titleObj = existingMetadata.get(Constants.KEY_TITLE);
+        if (titleObj == null) titleObj = currentPolicyDoc.get(Constants.KEY_TITLE);
+        mergedPolicy.setTitle(titleObj != null ? titleObj.toString() : "");
+
+        Object authorObj = existingMetadata.get(Constants.KEY_AUTHOR);
+        if (authorObj == null) authorObj = currentPolicyDoc.get(Constants.KEY_AUTHOR);
+        mergedPolicy.setAuthor(authorObj != null ? authorObj.toString() : "");
+
+        Object descObj = existingMetadata.get(Constants.KEY_DESCRIPTION);
+        if (descObj == null) descObj = currentPolicyDoc.get(Constants.KEY_DESCRIPTION);
+        mergedPolicy.setDescription(descObj != null ? descObj.toString() : "");
+
+        Object docObj = existingMetadata.get(Constants.KEY_DOCUMENTATION);
+        if (docObj == null) docObj = currentPolicyDoc.get(Constants.KEY_DOCUMENTATION);
+        mergedPolicy.setDocumentation(docObj != null ? docObj.toString() : "");
+
+        Object refObj = existingMetadata.get(Constants.KEY_REFERENCES);
+        if (refObj == null) refObj = currentPolicyDoc.get(Constants.KEY_REFERENCES);
+        @SuppressWarnings("unchecked")
+        List<String> existingReferences =
+                (List<String>) (refObj != null ? refObj : Collections.emptyList());
+        mergedPolicy.setReferences(existingReferences);
+
+        Object compatObj = existingMetadata.get(Constants.KEY_COMPATIBILITY);
+        if (compatObj == null) {
+            compatObj = currentPolicyDoc.get(Constants.KEY_COMPATIBILITY);
+        }
+        @SuppressWarnings("unchecked")
+        List<String> existingCompatibility =
+                (List<String>) (compatObj != null ? compatObj : Collections.emptyList());
+        mergedPolicy.getMetadata().setCompatibility(existingCompatibility);
+
+        Object existingRootDecoder = currentPolicyDoc.get("root_decoder");
+        String rootDecoderValue = existingRootDecoder != null ? existingRootDecoder.toString() : null;
+        mergedPolicy.setRootDecoder(
+                rootDecoderValue != null && !rootDecoderValue.isEmpty() ? rootDecoderValue : null);
+        mergedPolicy.setIntegrations(
+                (List<String>)
+                        currentPolicyDoc.getOrDefault(Constants.KEY_INTEGRATIONS, Collections.emptyList()));
+
+        // Apply the 5 modifiable fields from the incoming payload
+        mergedPolicy.setEnrichments(incomingPolicy.getEnrichments());
+        mergedPolicy.setFilters(incomingPolicy.getFilters());
+        mergedPolicy.setEnabled(incomingPolicy.getEnabled());
+        mergedPolicy.setIndexUnclassifiedEvents(incomingPolicy.getIndexUnclassifiedEvents());
+        mergedPolicy.setIndexDiscardedEvents(incomingPolicy.getIndexDiscardedEvents());
+
+        // Convert to JsonNode and persist.
+        // Ensure metadata fields are stored only under document.metadata.
+        ObjectNode policyNode = mapper.valueToTree(mergedPolicy);
+        Resource.nestMetadataFields(policyNode);
+
+        ContentIndex index = new ContentIndex(this.client, Constants.INDEX_POLICIES, null);
+        try {
+            ObjectNode document = mapper.createObjectNode();
+            document.set(Constants.KEY_DOCUMENT, policyNode);
+            ObjectNode spaceNode = mapper.createObjectNode();
+            spaceNode.put(Constants.KEY_NAME, Space.STANDARD.toString());
+            document.set(Constants.KEY_SPACE, spaceNode);
+            String hash = Resource.computeSha256(policyNode.toString());
+            ObjectNode hashNode = mapper.createObjectNode();
+            hashNode.put(Constants.KEY_SHA256, hash);
+            document.set(Constants.KEY_HASH, hashNode);
+            String standardPolicyId =
+                    this.spaceService.findDocumentId(
+                            Constants.INDEX_POLICIES, Space.STANDARD.toString(), docId);
+            IndexResponse indexResponse = index.create(standardPolicyId, document);
+            return indexResponse.getId();
+        } catch (Exception e) {
+            throw new IllegalStateException("Standard policy not found: " + e.getMessage());
+        }
+    }
+
+    /**
+     * Stores or updates the policy in the draft space.
+     *
+     * <p>If a draft policy already exists, it will be updated using its existing ID. Otherwise, a new
+     * policy will be created with a generated UUID.
+     *
+     * @param policy the policy to store
+     * @throws IOException if storage fails
+     */
+    @SuppressWarnings("unchecked")
+    private String updatePolicy(Policy policy) throws IOException, IllegalStateException {
+        // Get policy in the draft space
+        Map<String, Object> currentPolicy = this.spaceService.getPolicy(Space.DRAFT.toString());
+
+        // Safekeep unmodifiable values
+        Map<String, Object> currentPolicyDoc =
+                (Map<String, Object>) currentPolicy.get(Constants.KEY_DOCUMENT);
+        if (currentPolicyDoc == null) {
+            throw new IllegalStateException(
+                    Constants.E_500_INTERNAL_SERVER_ERROR + " Policy document not found in draft space.");
+        }
+
+        // Validate integrations: allow reordering but prevent addition/removal
+        List<String> currentIntegrations =
+                (List<String>)
+                        currentPolicyDoc.getOrDefault(Constants.KEY_INTEGRATIONS, Collections.emptyList());
+        List<String> newIntegrations = policy.getIntegrations();
+
+        // Validation for integrations array: allow reordering but prevent addition/removal
+        RestResponse validationError =
+                this.payloadValidations.validateListEquality(
+                        currentIntegrations, newIntegrations, Constants.KEY_INTEGRATIONS);
+        if (validationError != null) {
+            throw new IllegalArgumentException(validationError.getMessage());
+        }
+
+        // Validation for filters array: allow reordering but prevent addition/removal
+        List<String> currentFilters =
+                (List<String>)
+                        currentPolicyDoc.getOrDefault(Constants.KEY_FILTERS, Collections.emptyList());
+        List<String> newFilters = policy.getFilters();
+
+        RestResponse filtersValidationError =
+                this.payloadValidations.validateListEquality(
+                        currentFilters, newFilters, Constants.KEY_FILTERS);
+        if (filtersValidationError != null) {
+            throw new IllegalArgumentException(filtersValidationError.getMessage());
+        }
+
+        String docId = currentPolicyDoc.getOrDefault(Constants.KEY_ID, "").toString();
+        @SuppressWarnings("unchecked")
+        Map<String, Object> existingMeta =
+                (Map<String, Object>)
+                        currentPolicyDoc.getOrDefault(Constants.KEY_METADATA, Collections.emptyMap());
+
+        Object dateObj = existingMeta.get(Constants.KEY_DATE);
+        if (dateObj == null) dateObj = currentPolicyDoc.get(Constants.KEY_DATE);
+        String docCreationDate = dateObj != null ? dateObj.toString() : "";
+        String docModificationDate = Instant.now().toString();
+
+        // Update (set or overwrite unmodifiable values in incoming policy document)
+        policy.setId(docId);
+        policy.setDate(docCreationDate);
+        policy.setModified(docModificationDate);
+
+        Object compatObj = existingMeta.get(Constants.KEY_COMPATIBILITY);
+        if (compatObj == null) {
+            compatObj = currentPolicyDoc.get(Constants.KEY_COMPATIBILITY);
+        }
+        @SuppressWarnings("unchecked")
+        List<String> existingCompatibility =
+                (List<String>) (compatObj != null ? compatObj : Collections.emptyList());
+        policy.getMetadata().setCompatibility(existingCompatibility);
+
+        // Convert Policy to JsonNode.
+        // nestMetadataFields removes root-level duplicate fields (title, author, date, etc.)
+        // that Jackson emits from the public delegate getters in Policy, keeping them only
+        // inside the nested "metadata" object — consistent with how initializeSpace() works.
+        ObjectNode policyNode = mapper.valueToTree(policy);
+        Resource.nestMetadataFields(policyNode);
+
+        ContentIndex index = new ContentIndex(this.client, Constants.INDEX_POLICIES, null);
+        try {
+            // Build CTI wrapper with automatic hash calculation
+            ObjectNode document = mapper.createObjectNode();
+            document.set(Constants.KEY_DOCUMENT, policyNode);
+            ObjectNode spaceNode = mapper.createObjectNode();
+            spaceNode.put(Constants.KEY_NAME, Space.DRAFT.toString());
+            document.set(Constants.KEY_SPACE, spaceNode);
+            String hash = Resource.computeSha256(policyNode.toString());
+            ObjectNode hashNode = mapper.createObjectNode();
+            hashNode.put(Constants.KEY_SHA256, hash);
+            document.set(Constants.KEY_HASH, hashNode);
+            String draftPolicyId =
+                    this.spaceService.findDocumentId(Constants.INDEX_POLICIES, Space.DRAFT.toString(), docId);
+            IndexResponse indexResponse = index.create(draftPolicyId, document);
+            return indexResponse.getId();
+        } catch (Exception e) {
+            throw new IllegalStateException("Draft policy not found: " + e.getMessage());
+        }
+    }
+
+    /** Builds the engine payload for the standard space and loads it into the Engine. */
     private void loadStandardSpaceIntoEngine() {
         if (this.engineService == null) {
             log.warn(Constants.E_LOG_ENGINE_IS_NULL);
